The complete palette set with usability data
This file was autogenerated from data available from the the
ColorBrewer2 application SWF file, but with manual curation due to
differences between the palettes provided from the spreadsheet data and
the SWF - the spread sheet was taken to be the more accurate.
Additionally, there is a Paired palette with a further color added in the
spreadsheet palettes. The usability data for this additional palette was
extrapolated from the previous palette.

There still remain unresolvable ambiguities: in Greys and RdPu the L
color at 7 and 8 is the same as the K color at 9. This is at odds with
how the rest of the system seems to work. Rather than arbitrarily
picking a letter to assign, I have left it as it exists in the
spreadsheet. If further information becomes available that clarifies
this, this situation may change.